[CLEANUP] Migrate objects to proper namespaces
Part of #398
