+++ Classes/Mail.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use OliverKlee\Oelib\DataStructures\AbstractObjectWithAccessors;
 use OliverKlee\Oelib\Interfaces\MailRole;
 use Pelago\Emogrifier\CssInliner;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -13,7 +14,7 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  *
  * @author Niels Pardon <mail@niels-pardon.de>
  */
-class Tx_Oelib_Mail extends \Tx_Oelib_Object
+class Tx_Oelib_Mail extends AbstractObjectWithAccessors
 {
     /**
      * @var MailRole the sender of the e-mail
