+++ Tests/Unit/Email/Fixtures/TestingMailRole.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace OliverKlee\Oelib\Tests\Unit\Email\Fixtures;
 
+use OliverKlee\Oelib\DataStructures\AbstractObjectWithAccessors;
 use OliverKlee\Oelib\Interfaces\MailRole;
 
 /**
@@ -11,7 +12,7 @@ use OliverKlee\Oelib\Interfaces\MailRole;
  *
  * @author Niels Pardon <mail@niels-pardon.de>
  */
-class TestingMailRole extends \Tx_Oelib_Object implements MailRole
+class TestingMailRole extends AbstractObjectWithAccessors implements MailRole
 {
     /**
      * @var string[] the data of this object
